Make a "trampoline" crate to avoid unneeded dependencies in a proc-macro crate
